pyproject.toml
[tool.poetry]
name = "softdesk"
version = "0.1.0"
description = "Api for project management"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.7.2"
django = "4.2.5"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.3.0"
pyjwt = "2.8.0"
pytz = "2023.3.post1"
sqlparse = "0.4.4"
tzdata = "2023.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
